Refactor model manager inference api
